Abstract out check_reserved_keyword
- Add utils.py
- Delete _exceptions.py; Move NoException to utils.py
