--- src/elapi/_exceptions.py
@@ -1,1 +1,0 @@
-class NoException(Exception): ...
